LRQA-43331 Move writing properties files into WorkspaceGitRepository
